backend/requirements/development.txt
-r base.txt
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.27.0
black==24.8.0
flake8==7.1.1
isort==5.13.2